rename all Foo/Type to Foo/Foo
this commit was auto-generated and might break things. if it breaks *too
many* things, feel free to revert it. this change is a small polish that
simplifies the logic of finding the file where a term resides, since now
we don't ever consider that "Foo/Bar" might be in "Foo/Bar/Type", which
was ugly and hard-coded. this changes keeps Kind slightly simpler

I've also added some Kind files to this repo